Move vue components to separate directory
